requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
Django==5.0.1
django-crispy-forms==2.1
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
